+++ CHANGELOG.md
@@ -14,6 +14,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [3.0.1] - Upcoming
 
+### Changed
+
+- Updated private methods to static in `InteractsWithApi` trait.
+
 ### Fixed
 
 - Bugfix in `getBody` method returning `null`.